\docType{data}
\name{demoList1}
\alias{demoList1}
\alias{demoList2}
\title{DAVID's website demoList1 example id files}
\format{character vector with AFFYMETRIX_3PRIME_IVT_ID manufacturer
identification codes (ids)
\describe{
\item{demoList1}{164 ids in total.}
\item{demoList2}{403 ids in total.}
}}
\usage{
  data(demoList1)

  data(demoList2)
}
\description{
  This datasets are the same example input id files present
  in the Database for Annotation, Visualization and
  Integrated Discovery.
}
\author{
  Cristobal Fresno and Elmer A Fernandez
}
\references{
  \enumerate{ \item The Database for Annotation,
  Visualization and Integrated Discovery
  (\url{david.abcc.ncifcrf.gov}) \item Huang, D. W.;
  Sherman, B. T.; Tan, Q.; Kir, J.; Liu, D.; Bryant, D.;
  Guo, Y.; Stephens, R.; Baseler, M. W.; Lane, H. C.;
  Lempicki, R. A. DAVID Bioinformatics Resources: expanded
  annotation database and novel algorithms to better
  extract biology from large gene lists. Nucleic Acids Res,
  Laboratory of Immunopathogenesis and Bioinformatics,
  SAIC-Frederick, Inc., National Cancer Institute at
  Frederick, MD 21702, USA., 2007, 35, W169-W175 }
}
\seealso{
  Other DataExamples: \code{\link{annotationSummary1}},
  \code{\link{annotationSummary2}},
  \code{\link{geneList1}}, \code{\link{geneList2}}
}
\keyword{datasets}

