\docType{methods}
\name{type}
\alias{type,DAVIDResult-method}
\alias{type}
\title{Getters for \code{DAVIDResult} object}
\usage{
  \S4method{type}{DAVIDResult}(object)
}
\arguments{
  \item{object}{DAVIDResult class object.}
}
\value{
  according to the call one of the following objects can be
  returned \item{type}{character with type slot datum.}
}
\description{
  Obtain DAVIDResult slot information, according to the
  given function call (see values).
}
\author{
  Cristobal Fresno and Elmer A Fernandez
}
\seealso{
  Other DAVIDResult: \code{\link{DAVIDResult-class}},
  \code{\link{plot2D}}, \code{\link{plot2D}},
  \code{\link{plot2D}}, \code{\link{plot2D}},
  \code{\link{plot2D}}, \code{\link{plot2D}}
}

