\docType{package}
\name{DAVIDWebService-package}
\alias{DAVIDWebService-package}
\alias{DAVIDWebService-package}
\title{An R Package for retrieving data from DAVID into R objects using Web Services}
\description{
  Tools for retrieving data from the Database for
  Annotation, Visualization and Integrated Discovery
  (DAVID) using Web Services into R objects. This package
  offers the main functionalities of DAVID website
  including: \describe{ \item{Connectivity:}{upload
  gene/background list/s, change gene/background position,
  select current specie/s, select annotations, etc. from
  R.} \item{Exploration:}{ native R objects of submitted
  Gene List, Annotation Category Summary, Gene/Term
  Clusters, Functional Annotation Chart and Functional
  Annotation Tables. In addition it offers the usual
  many-genes-to-many-terms visualization and induced Gene
  Ontology direct acyclic graph GOstats-based conversion
  method, in order to visualize GO structure.} }
}
\author{
  Cristobal Fresno \email{cristobalfresno@gmail.com} and
  Elmer A. Fernandez \email{elmerfer@gmail.com}
}
\references{
  \enumerate{ \item The Database for Annotation,
  Visualization and Integrated Discovery
  (david.abcc.ncifcrf.gov) \item Xiaoli Jiao, Brad T.
  Sherman, Da Wei Huang, Robert Stephens, Michael W.
  Baseler, H. Clifford Lane, Richard A. Lempicki, DAVID-WS:
  A Stateful Web Service to Facilitate Gene/Protein List
  Analysis Bioinformatics 2012
  \url{doi:10.1093/bioinformatics/bts251} \item Huang, D.
  W.; Sherman, B. T.; Tan, Q.; Kir, J.; Liu, D.; Bryant,
  D.; Guo, Y.; Stephens, R.; Baseler, M. W.; Lane, H. C. &
  Lempicki, R. A DAVID Bioinformatics Resources: expanded
  annotation database and novel algorithms to better
  extract biology from large gene lists. Nucleic Acids Res,
  Laboratory of Immunopathogenesis and Bioinformatics,
  SAIC-Frederick, Inc., National Cancer Institute at
  Frederick, MD 21702, USA., 2007, 35, W169-W175 
  \item Cristobal Fresno, Elmer A. Fernandez (2013) 
  RDAVIDWebService: a versatile R interface to DAVID, 
  Bioinformatics, 29(21), 2810-2811.,
  \url{http://bioinformatics.oxfordjournals.org/content/29/21/2810.}}
}
\keyword{DAVID}
\keyword{MEA}
\keyword{SEA}

